+++ types/changelog.md
@@ -28,6 +28,8 @@ All notable changes to this project will be documented in this file.
   ([#121](https://github.com/kotools/libraries/issues/121)).
 - The `StrictlyPositiveInt.toPositiveInt` function
   ([#122](https://github.com/kotools/libraries/issues/122)).
+- The `StrictlyNegativeInt.toNegativeInt` function
+  ([#123](https://github.com/kotools/libraries/issues/123)).
 
 #### Removed